requirements.txt
Flask
PyYAML
gpsoauth>=0.4.1
six>=1.10.0
future>=0.16.0
enum34>=1.1.6; python_version < '3.0'